api_domain setting for non-primary domain requests
Due to the massive DDoS, some of Linode's API endpoints do not work properly. The suggested fix was to use one of a set of IPs for the API. If we make a request against the IP instead of domain, we need to pass the domain in the "host" header. This allows us to optionally set that header.